refactor: split architectural optimisation tests into module-focused files
- WS-1 pagination tests → query-sparql.test.ts
- WS-2 deepQuery/getter tests → Ad4mModel.test.ts
- WS-7 proxy cache tests → PerspectiveClient.test.ts (new)
- Delete monolithic architectural-optimisation.test.ts
- 412/414 tests passing (1 pre-existing failure in Ad4mClient.test.ts)
